test: a user can not delete himself

+++ medical-appointment-app-26-1-8CS/app/Http/Controllers/Admin/UserController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Spatie\Permission\Models\Role;
+use illuminate\Support\Facades\Auth;
 
 class UserController extends Controller
 {
@@ -109,6 +110,19 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+
+        //No permitir que el usuario loguado se borre a sí mismo
+        if ($user->id == Auth::user()->id) {
+            session()->flash('swal', [
+                'icon' => 'error',
+                'title' => 'Accion denegada',
+                'text' => 'No puedes eliminar tu propia cuenta.'
+            ]);
+            abort(403,'No puedes borrar tu propio usuario');
+
+            return redirect(route('admin.users.index'));
+        }
+        
         //Eliminar roles asociados a un usuario
         $user->roles()->detach();
